Move custom exceptions and global vars to dedicated files, add try catch for all CLI commands
